config file

+++ install.sh
@@ -4,7 +4,18 @@ echo "---cp tailwin.desktop /usr/share/xsessions---"
 sudo cp tailwin.desktop /usr/share/xsessions
 echo "---cp tailwin.png /usr/share/pixmaps---"
 sudo cp tailwin.png /usr/share/pixmaps
-echo "---cargo build --release---"
-cargo build --release
-echo "---ln -f ./target/release/tailwin /usr/bin/tailwin---"
-sudo ln -f ./target/release/Tailwin /usr/bin/tailwin
+echo "---if file exists---"
+FILE=./src/tailwin.rs
+if [ -f "$FILE" ]; then
+    echo "---cargo build --release---"
+    cargo build --release
+    echo "---ln -f ./target/release/tailwin /usr/bin/tailwin---"
+    sudo ln -f ./target/release/Tailwin /usr/bin/tailwin
+else 
+    echo "---mkdir ~/.config/tailwin---"
+    mkdir ~/.config/tailwin
+    echo "---printf 'fn on_startup() {\n\t// Do stuff\n}' > ~/.config/tailwin/tailwin.rs---"
+    printf "pub fn on_startup() {\n\t// Do stuff\n}" > ~/.config/tailwin/tailwin.rs
+    echo "---ln -f ~/.config/tailwin/tailwin.rs ./src/tailwin.rs---"
+    ln -f ~/.config/tailwin/tailwin.rs ./src/tailwin.rs
+fi
